Revert "rename structure test directory for consistency (#365)"
This reverts commit e70bea4dcdc2b1cc4327a752f3d1dd261455ae00.
