+++ structure_tests/drivers/driver.go
@@ -18,7 +18,7 @@ import (
 	"os"
 	"testing"
 
-	"github.com/GoogleCloudPlatform/runtimes-common/structure-test/types/unversioned"
+	"github.com/GoogleCloudPlatform/runtimes-common/structure_tests/types/unversioned"
 )
 
 type Driver interface {
