+++ structure_tests/ext_run.sh
