+++ structure_tests/run.sh
