+++ structure_tests/structure_test.go
@@ -25,7 +25,7 @@ import (
 	"strings"
 	"testing"
 
-	"github.com/GoogleCloudPlatform/runtimes-common/structure-test/drivers"
+	"github.com/GoogleCloudPlatform/runtimes-common/structure_tests/drivers"
 	"github.com/ghodss/yaml"
 )
