+++ structure_tests/tests.bzl
