+++ structure_tests/types.go
@@ -17,9 +17,9 @@ package main
 import (
 	"testing"
 
-	"github.com/GoogleCloudPlatform/runtimes-common/structure-test/drivers"
-	"github.com/GoogleCloudPlatform/runtimes-common/structure-test/types/v1"
-	"github.com/GoogleCloudPlatform/runtimes-common/structure-test/types/v2"
+	"github.com/GoogleCloudPlatform/runtimes-common/structure_tests/drivers"
+	"github.com/GoogleCloudPlatform/runtimes-common/structure_tests/types/v1"
+	"github.com/GoogleCloudPlatform/runtimes-common/structure_tests/types/v2"
 )
 
 type StructureTest interface {
