+++ structure_tests/types/v1/file_content.go
