+++ structure_tests/types/v1/licenses.go
@@ -20,7 +20,7 @@ import (
 	"strings"
 	"testing"
 
-	"github.com/GoogleCloudPlatform/runtimes-common/structure-test/drivers"
+	"github.com/GoogleCloudPlatform/runtimes-common/structure_tests/drivers"
 )
 
 // Not currently used, but leaving the possibility open
