+++ structure_tests/types/v2/file_content.go
