+++ structure_tests/utils/utils.go
